adapt and requite libowfat 0.30
The signature of dns_domain_fromdot() changed between 0.29 and 0.30. Adapt to
the new one, which in turn dropy support for the older versions. Since it's
only statically linked everyone can freely update the library locally before
upgrading Qsmtp.

+++ tests/qdns_dane_test.c
@@ -181,7 +181,7 @@ static const char success_packet[] = "\0\0\0\0\0\0\0\3\0\0\0\0" /* header */
 		"\x40\x41\x42\x43\x44\x45\x46\x47\x48\x49\x4a\x4b\x4c\x4d\x4e\x4f"; /* TLSA data (SHA2-512) */
 
 int
-dns_domain_fromdot(char **q, const char *host, unsigned int len)
+dns_domain_fromdot(char **q, const char *host, size_t len)
 {
 	unsigned int idx;
